--- src/apply/text/index.js
@@ -1,9 +1,0 @@
-const insertText = require('./insertText');
-const removeText = require('./removeText');
-
-const mappers = {
-    insert_text: insertText,
-    remove_text: removeText,
-};
-
-module.exports = mappers;
