+++ src/apply/text/index.ts
@@ -1,0 +1,7 @@
+import insertText from './insertText';
+import removeText from './removeText';
+
+export default {
+    insert_text: insertText,
+    remove_text: removeText,
+};
